* develop: (33 commits)
  fix black screen when back to foreground  on android.
  Fix set FPS broken.
  Update CHANGELOG [ci skip]
  Update AUTHORS [ci skip]
  Update CCString.cpp
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Updates Bindings-generator.
  Renames: IntValueMap —> ValueMapIntKey
  Restricts the element type for Vector<T> and Map<K, V>, it has to be `Object*` or `ObjectSubClass*`.
  Modify reader
  long --> ssize_t for TableView callback.
  Generates execute file in cocos root_dir/bin folder for HelloCpp.
  Add listview and pageview  reader
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Modify actionnode
  [AUTO] : updating submodule reference to latest autogenerated bindings
  optimize loop
  fix LabelTestNew
  Compilation fixes after merge PR #4669 https://github.com/cocos2d/cocos2d-x/pull/4669
  Fix box2dView
  ...

+++ samples/Cpp/TestCpp/Classes/ExtensionsTest/TableViewTest/TableViewTestScene.cpp
@@ -96,7 +96,7 @@ TableViewCell* TableViewTestLayer::tableCellAtIndex(TableView *table, ssize_t id
     return cell;
 }
 
-long TableViewTestLayer::numberOfCellsInTableView(TableView *table)
+ssize_t TableViewTestLayer::numberOfCellsInTableView(TableView *table)
 {
     return 20;
 }
